Add base branch matching
Allow the specifying of the base branch along with the matching of a
head branch. This is a quite a simple change as we use the same code for
branch matching, but tailor it for base branches.

This is great for my usage, as I can then label new PRs dependant on
where they're going to land, helping priortize and categorize incoming
changes to a repo.

Yaml naming