Add Target to abstract a PR/Issue
According to the google/go-github docs, every PR is also an Issue (but
not in reverse). We could duck-type our way into treating both of them
equally, but in order to evaluate conditions properly we do need access
to data that is not available behind the same functions (e.g. RepoName),
as well as the raw PR/Issue.

To work around this, we rewrite all the main functions to rely on a new
Target struct, that abstracts away the differences between PR and Issue,
while giving access to both.

Conditions that rely on a PR make sure that the Target does contain a
PR, or else they evaluate to false.

[1]: https://pkg.go.dev/github.com/google/go-github/v35/github#Issue

Signed-off-by: Galo Navarro <anglorvaroa@gmail.com>