size: composite configuration
In retrospect, the previous config for size was not very fortunate as it
splits properties that have a semantic relation in two unrelated ones:

    - label: M
        size-above: 10
        size-below: 100

In order to support file exclusions for the size calculation it'd make a
cleaner config schema to have all size properties in a single Size node.

This change introduces a `size` node as a condition configuration that
replaces `size-above` and `size-below`, while respecting backwards
compatibility. Older config files will continue working, but from now on
the way to specify the file condition will be:

        - label: M
            above: 10
            below: 100

Signed-off-by: Galo Navarro <anglorvaroa@gmail.com>